Move sign, verify to appropriate namespace
